docs(rfc-0005): RFC 0005 r1 — reconcilia com RFC 0003 implementada em JS
