+++ package.json
@@ -17,22 +17,22 @@
     "astro": "astro",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "hronir": "node scripts/hronir/index.js",
-    "hronir:next": "node scripts/hronir/index.js next",
-    "hronir:init": "node scripts/hronir/index.js init",
-    "hronir:continue": "node scripts/hronir/index.js continue",
-    "hronir:decide": "node scripts/hronir/index.js decide",
-    "hronir:ranking": "node scripts/hronir/index.js ranking",
-    "hronir:worst": "node scripts/hronir/index.js worst",
-    "hronir:diagnose": "node scripts/hronir/index.js diagnose",
-    "hronir:edit-worst": "node scripts/hronir/index.js edit-worst",
-    "hronir:edit-commit": "node scripts/hronir/index.js edit-commit",
-    "hronir:draft-worst": "node scripts/hronir/index.js draft-worst",
-    "hronir:draft-commit": "node scripts/hronir/index.js draft-commit",
-    "hronir:promote": "node scripts/hronir/index.js promote",
-    "hronir:end": "node scripts/hronir/index.js end",
-    "hronir:migrate": "node scripts/hronir/index.js migrate",
-    "hronir:doctor": "node scripts/hronir/index.js doctor",
+    "hronir": "node --import tsx/esm scripts/hronir/index.js",
+    "hronir:next": "node --import tsx/esm scripts/hronir/index.js next",
+    "hronir:init": "node --import tsx/esm scripts/hronir/index.js init",
+    "hronir:continue": "node --import tsx/esm scripts/hronir/index.js continue",
+    "hronir:decide": "node --import tsx/esm scripts/hronir/index.js decide",
+    "hronir:ranking": "node --import tsx/esm scripts/hronir/index.js ranking",
+    "hronir:worst": "node --import tsx/esm scripts/hronir/index.js worst",
+    "hronir:diagnose": "node --import tsx/esm scripts/hronir/index.js diagnose",
+    "hronir:edit-worst": "node --import tsx/esm scripts/hronir/index.js edit-worst",
+    "hronir:edit-commit": "node --import tsx/esm scripts/hronir/index.js edit-commit",
+    "hronir:draft-worst": "node --import tsx/esm scripts/hronir/index.js draft-worst",
+    "hronir:draft-commit": "node --import tsx/esm scripts/hronir/index.js draft-commit",
+    "hronir:promote": "node --import tsx/esm scripts/hronir/index.js promote",
+    "hronir:end": "node --import tsx/esm scripts/hronir/index.js end",
+    "hronir:migrate": "node --import tsx/esm scripts/hronir/index.js migrate",
+    "hronir:doctor": "node --import tsx/esm scripts/hronir/index.js doctor",
     "music:generate": "node scripts/generate-music-posts.mjs",
     "test": "node --experimental-test-module-mocks --test scripts/hronir/lib/__tests__/ranking.test.js"
   },
@@ -65,6 +65,7 @@
     "openskill": "^4.1.1",
     "playwright": "^1.60.0",
     "prettier": "^3.8.3",
-    "remark": "15.0.1"
+    "remark": "15.0.1",
+    "tsx": "^4.22.4"
   }
 }
